Add internal json/jsonb support to flex output
Table columns with type 'json' or 'jsonb' are now explicitly supported.
Lua booleans, strings, numbers, and tables can be written to those
columns and are automatically encoded as JSON on the fly.

This is a breaking change to the behaviour we had before where you had
to encode the JSON yourself in Lua before giving it to osm2pgsql. We
can still do this change as the flex backend is still marked
experimental and the behaviour is much cleaner this way then if we keep
backwards compatibility.