Refactor middle interface to get relation members
In the future we want the middle to be able to return not only way
members of relations but any type of member. This renames the
function rel_way_members_get() to rel_members_get() and adds an extra
parameter that specifies which types of members we are interested
in. The ram middle already supports accessing non-way members, but
the pgsql middle still only supports way members.