Only create create_only columns in final table
The flex output supports "create_only" table columns which are only
created but not filled with any data. This can lead to problems in some
cases (for instance when a column of type SERIAL is used) because
osm2pgsql usually creates tables twice, first a temporary table, then a
final one. This commit changes the behaviour slightly: The "create_only"
column is only created in the final table.

Fixes #1522