regression tests: fix usage of psycopg2.connect
Contrary to what one might expect, the connection context manager
does not close the connection when leaving but just end the
currently running transaction.

This commit removes all uses of the 'with' construct with connections
and replaces them with a classic try/finally block, explicitly
closing the connection on the way.

Fixes a breakage of the tests when used with the psycopg2 2.9
release.