Clean up pointers to middle_query_t in outputs to release memory
The outputs have a shared pointer to the middle_query_t to access the
data held by the middle. This in turn holds pointers to potentially
large datastructures and the flat node file. We need to release those
pointers after we don't need them any more so that if and when the
middle doesn't need those data structures any more they get cleaned up.

In effect this releases the memory used for the cache and flat node
store earlier.

Fixes #1563