Large refactoring of the geometry code
Pulls apart the different concerns:
* geometry generation from OSM data
* geometry processing (like merging and splitting linestrings)
* conversion to WKB

For all of this to work we switch to C++17 as minimum requirement.

This also adds some code (in src/geom-boost-adaptor.hpp) that adapts our
data structures to boost::geometry. This is not used yet, but can in the
future be used for more complex geometry processing.