Fix parsing problems in style file reader
Some variables were not initialized correctly when parsing a style file,
which lead to some surprising behaviour with flags of one config line
re-used by the next if the flags field of that line was empty.

This could also have lead to buffer overflows in the first line being
parsed.

This commit also adds some more tests for the parsing code.

See #1590