Work-in-progress: Add geometry functions in Lua / new insert syntax
This is the future of geometry processing in Lua. Only the most
important geometry functions are implemented so far.

There is a new syntax for inserting rows into a table. Instead of
"add_row" the "insert" function is used. "add_row" is still available
and it works as before. But the new geometry functions only work with
the "insert" command. The "insert" function also allows multiple
geometry columns which wasn't possible with "add_row".

Geometry creation functions:
* as_point() (for nodes)
* as_linestring() (for ways)
* as_polygon() (for ways)
* as_multilinestring() (for relations)
* as_multipolygon() (for relations)
* as_geometrycollection() (for relations)

Geometry functions:
* geometrytype()
* srid()
* transform(target_srid)
* area() (for polygons)
* centroid() (for polygons)
* simplify(tolerance) (for linestrings)
* split_multi() (for multi geometries)
* #geom/__len() returns the number of geometries (0 for null geom, 1 for
  point/linestring/polygon, n for multi* and geometrycollection geoms)

See the file flex-config/new-insert-syntax.lua for an example that shows
how to use these.