Add support for the get_bbox() function for relations
The get_bbox() function now works for relations, too. Only node and way
members are taken into account. The data is cached properly so that it
is only requested from the database once even if get_bbox() and/or
add_row() are called multiple times.

Part of this change is that we'll get the location of all nodes members
of relations. This is new, we used to only get the way members and the
locations in those ways. Getting all the node members keeps the code
simpler, but has some overhead. It should be fairly small though
compared to getting all the way members of the multipolygon and boundary
relations which make up the bulk of what we are using.