Do not reuse Osmium buffer
This cleans up a bit of confusing code. We were reusing the m_buffer
member in two places where a local Buffer object is fine. The idea
was to save a memory allocation here for the local Buffer, but is
probably not worth it and muddies up what the m_buffer is actually
used for.