Add dimension() function for geometry types
This adds a dimension() function for all geometry types. The topological
dimension is defined as 0 for point geometries, 1 for linear geometries
and 2 for areal geometries. For geometry collections it is the largest
dimension of any of its members.

See https://postgis.net/docs/manual-3.3/ST_Dimension.html

It is needed for the implementation of centroid() for geometry
collections.