Add get_node_location() function to middle
Before we had only a function to retrieve a list of locations from a
list of nodes which is inconvenient if you only need a single location.

The new function is needed in the context of retrieving locations for
member nodes of relations.

There are probably ways to improve on this:
* We are using the get_node_list prepared statement to get the single
  id when a simpler prepared statement would do. But then we'd need
  another prepared statement.
* When using this it might be possible to first assemble a list of
  several ids and get them all in one go. But its more difficult to
  write code like this and there aren't that many relations around
  that have lots of node members, so we'll leave it at that and
  optimize this later if needed.