Use extra functions for each type to be moved into geometry_t
The template we used before interfered with the normal operation of the
class because it matched geometry_t itself, so it worked as copy
constructor. With the explicit functions we remove that problem and make
sure the different types can only be moved into the geometry_t object,
so we don't copy unnecessarily without noticing.