Ensure that flex_table_t::geom_column() refers to the first one
Now that there are multiple geometry columns allowed this makes sure we
always have the first one (which is important for clustering).

Remove the `column.set_not_null();` in `add_column()` for geometry
columns because it did not work anyway. It is overwritten by the
`column.set_not_null(...)` call right outside the `add_column()` call.