Add pole_of_inaccessibility() function in Lua for polygons
Implementation of the "polylabel" algorithm for finding the "pole of
inaccessibility" or center of the maximum inscribed circle of a polygon.
This is accessible from Lua as the pole_of_inaccessibility() function.

The current implementation is only defined for polygons, because it is
unclear how it should behave for other types of geometries. We can
extend this later if needed.