Fix: Create point geometry from collection with single-node way
If a geometry collection is created from a relation, any way member will
normally end up as linestring relation. But if there is only a single
point in the resulting linestring (because the way only contained a
single node or multiple nodes with the same position), that linestring
would be invalid. In this case create a point geometry instead.