Refactor of expire code: Move tile writing code out of expire_tiles
In the future we want to have different ways of reporting expired
tiles. This splits out the code to write tiles to a file to make
it more straightforward to add other methods in the future.

The tests are also a bit more straightforward now and some now also test
the order the tiles are reported in.

This also adds a file tile-output.hpp which contains a streams
implementation for writing out tiles. Used for debugging when tests
fail.