Refactor expire by geometry code, disable expire when not using 3857
Split up code into functions for each geometry type. Adds support for
geometry types that have been missing so far (multipoint, geometry
collection).

When using a target SRS other than Mercator (EPSG:3857) expire didn't
work before, it is now explicitly disabled (with a warning) in this
case. (This applies only to the pgsql output, the flex output sets
the SRS per table. There is currently no warning in this case, but
we are moving away from the expire configuration on the command line
for the flex output anyway, so I'll leave it at that for the moment.)