Add tilesets to flex output allowing more flexible expire configuration
Tilesets are defined with `define_tileset()` in the Lua config file.
They have a name, min and max zoom levels, and a filename or table name
(optionally with schema). The filename or table specify where the
contents of the tilesets are written to at the end of an osm2pgsql run.

Tilesets are referenced from geometry columns of database tables. A new
`expire` field allows setting any number of tilesets and some additional
parameters. In append mode, any geometry written to those columns
triggers an entry in the corresponding tileset(s). Deletes also trigger
these entries.

For backwards compatibility an expire config set with -e/--expire-tiles
and -o/--expire-output is automatically added to all tables. Those
command line options (together with --expire-bbox-size) should be marked
as deprecated at some point and later removed.