WIP: Add code to generalize data using different strategies
This commit adds the code to generalize various types of data using
different strategies.

The following strategies work on a tile-by-tile basis and operate on
polygons:

The "vector-union" strategy buffers and unionizes polygons using vector
operations.

The "raster-union" strategy does a similar thing but does it in raster
space which is much faster. First the polygons are rendered into a
raster, an open/close operation is called (which basically does the same
thing as the buffering in vector space) and finally the resulting raster
is vectorized again.

The "builtup" strategy is intended to derive a layer of builtup areas
from landuse=residential/industrial etc. as well as building cover and
dense road networks. This still needs some work...

Also a new "discrete-isolation" strategy which rates places based on
some importance metric. (This is not tile-based.)

The new "rivers" strategy finds important rivers, this is still very
much work in progress.

For the raster support this adds two new library dependency: CImg and
potrace.

The functionality is accessed through a new command line program called
osm2pgsql-gen. It reads the same Lua config file that osm2pgsql reads.
Call it with -h to get some usage information. This program is for
testing only, eventually the functionality should be accessible from
osm2pgsql itself.

See also https://osm2pgsql.org/generalization/ .