Make timer_t use microseconds and make it restartable
We can use the timer in more places if we have better resolution. For
places where we don't need the resolution it doesn't matter because it
will just get rounded. While we are at it we also use microseconds in
the thread_pool code so that we use the same unit everywhere.

The timer is restartable now, by repeatedly calling start/stop() we can
add up the times.

Also we are trying, as much as possible, to keep to the time types
supplied by chrono. This gives us automatic correct time output with the
fmt library which knows about those types. And it should make time
conversions easier.