Force database client encoding to UTF8
Doesn't hurt and might prevent some problems, because osm2pgsql is
always dealing with UTF8 data.

Note that, just as with the fallback_application_name, this setting
isn't set in all cases, but we don't want to meddle with the user
settings too much to not break the connection info.