Remove pg_conn_t::query() function, use pg_conn_t::exec() instead
This removes the pg_conn_t::query() in favour of the pg_conn_t::exec()
function. They mostly did the same thing anyway and it was always
confusing when to use which one and how to set the "expect" parameter.

Now we have exec() and exec_prepared() as the only functions running a
SQL command and returning the results.