Use fmt::format() function instead of ""_format literal
This replaces most uses of the ""_format literal with function calls to
fmt::format(). The latter form is available in all versions of the fmt
library, while the ""_format literal is deprecated in version 8 and
removed in version 9.

This might cost us some performance and early error detection, because
in at least some versions of the fmt library the _format() construct was
used to compile the format patterns at run time. That's why there are
some uses of the literal left in places where performance might be more
critical, they need a closer look.

See #1704