Remove the last uses of _format()
Three of the four cases are simple, we can use fmt::format and don't
need to parse a format string. The last case has the format string
R"(\u{:04x})" to parse, but that's only needed when some control
characters are written out in JSON format, which doesn't happen that
often.

Fixes #1704