Simplify db_copy_mgr_t::add_value()
The reason we had a different implementation for "double" values here is
that std::to_string() uses the current locale for writing out numbers.
For integers this doesn't matter, but for doubles this could be
something different than a decimal point. The fmt::to_string() function
on the other hand always uses the default locale, so it doesn't have
this problem. So we can use this for integers as well as for doubles.