Add and use expire_config_t struct
This commit introduces a new struct expire_config_t which collects some
configuration options for expire. All functions that calculate which
tiles to expire from some geometry now get an instance of this struct
as parameter.

Config parameters are:
* `buffer` decides how far around the actual geometries the tiles
  should be expired. This replaces the "tile_expiry_leeway" constant.
  It is always set to 0.1 (= 10% of tile size) as before and can
  currently not be changed.
* `boundary_only` is a new setting. If this is set, polygons are
  not expired by bounding box, but only along the boundary. Currently
  always set to false and can not be changed.
* `max_bbox` used to be a setting of the same name in the expire_tiles
  class. It is set by the `--expire-bbox-size=SIZE` option. It specifies
  the max width and height for a polygon to expire the whole polygon,
  not just the boundary (default 20000). If `boundary_only` is false,
  this setting can still overwrite that. The function can be disabled
  by setting it to 0.0 (this is new, it could not be completely
  disabled before).

From here on the expire_tiles class is only responsible for actually
keeping a list of expired tiles while the config which tiles are to
be expired is added dynamically when the actual expiry happens. This
will allow us in the future to have different layers with different
expiry options feed into the same expired tile list.