Add option to always build id index
If the user always needs the id index, they can force the index build
even in non-slim (or slim+drop) mode by setting the `create_index`
option of the `ids` setting in the define_table() Lua command to
`always`. The default is `autpo` which means: Only build the index in
slim mode.

See #1854

The test runner (steps_db.py) has been extended so that the
table_exists() check also checks for views, so that we can access the
data in the pg_catalog.pg_indexes view.