Flush COPY between nodes/ways and ways/relations respectively
In most cases database tables are filled based on OSM object types, so
after we have read all nodes nothing will be added to node-type tables
any more etc. So it makes sense to explicitly flush the COPY making sure
this data is written to the database. If more data is later added, maybe
because of second-stage processing, it doesn't matter, the COPY will be
reopened.