Merge the flex_tileset_t class into expire_tiles class
Using two different classes was an attempt at splitting up configuration
for the flex output and the actual functionality of storing the tile
coordinates (used in flex and pgsql output), but this turns out to make
the code more complex than it needs to be.

Note that there is still the code in output-flex.[ch]pp and
flex-lua-tileset.[ch]pp which talks about tilesets, that's also the name
that the end user sees. We still need to decide what to call this thing
and straighten this out.