Move initialization of thread number into logger code
This encapsulates the thread local variable this_thread_num.

We also set the thread name in the same place. The naming of the thread
has changed because Linux only allows 16 characters for the thread name
(including the \0), so before the actual name only ever came out as
"_osm2pgsql_work".