Add functionality to send data to PostgreSQL in binary
Prepared statements can now send parameters in binary as well as in the
default text format. To do this, wrap the binary data in an instance of
the binary_param class, which is just a wrapper of std::string_view used
to signal to the exec_prepared() function that this parameter should be
sent in binary.

You can mix and match any number of parameters in text or binary format.

It isn't possible any more to use exec_prepared() without any
parameters, but that wasn't use (and doesn't make much sense anyway), so
no loss there.