Store properties to a database table and retrieve for updates
Store command line settings in a database table called
"osm2pgsql_properties" when importing data. This is mainly so that we can
do updates without specifying those settings again (or checking that the
settings match the ones used for import.)

The table always has the same name and will be stored in the schema
used for the middle tables (set with --middle-schema).

Adds a new generic properties_t class which can store strings, integers
and booleans as properties. Internally a map of strings is used for the
properties, in the database a simple table with two text columns,
property and value, is used. For boolean values strings 'true' and
'false' are used.

Currently the following properties are stored in the database:
- "attributes" (bool): -x/--extra-attributes setting
- "flat_node_file": flat node file name
- "prefix": prefix from --prefix
- "updatable" (bool): can the database be updated, i.e. created with
  --slim and no --drop
- "version": osm2pgsql version