Add properties from input file(s) to database
This adds two new properties to the database: "import_timestamp" and
"current_timestamp". Both initially contain the highest timestamp found
on any of the objects in the input file(s). The "import_timestamp" will
never be changed again, the "current_timestamp" will be updated in
append mode. If the input file(s) don't contain timestamps, the
properties will not be created.

It also adds properties "replication_base_url",
"replication_sequence_number", and "replication_timestamp" if there is
only a single input file and if that file has the respective header
fields set.