Refactor middle: Pull special osm_* tags out of the tag list
The middle stores OSM attributes (version, timestamp, changeset, uid,
user) in "osm_*" pseudo-tags. When we get data back from the middle, we
used to just leave them in there. This commit changes that behaviour: We
pull the data out and put it into the proper attributes and remove the
special tags. This is how the flex output needs the data and this is how
a new middle (we are currently working on) will provide the data.

The pgsql output puts the pseudo-tags in already, because it had to do
that on import anyway.

This also fixes a problem where the "osm_*" pseudo-tags were visible in
the Lua config file of the flex output.