Order tags in input files
This is how they would appear in real data and it is better to be
consistent to avoid confusion. It doesn't matter for the test here
currently, but in the future tags will be stored differently in the
database and need to be re-ordered after reading from the database and
then this test will fail.