Clean up lua includes
There was a bit of a mess of different (types of) includes. We now only
use '#include <lua.hpp>' which includes all Lua C headers correctly and
also works for LuaJIT.

Technically it might include more than we need in some cases, because
lua.hpp includes several C headers some of which are only needed
sometimes, but the largest one (lua.h) is the one we always need, the
others are tiny.