Add special id column type "tile" for generalized data
Generalized data is often stored in tables accessed by tile (x/y). This
adds special code for that so we don't have to defined these columns
explicitly in the Lua config. The columns are logically very similar to
the "id column(s)" stored for "normal" output tables in that they are
used by osm2pgsql to update the data correctly. So we are using the same
mechanism for these columns.

In "create" mode an index is automatically generated on these columns
after generalization. As with the other id indexes this will happen
either when the database is updatable or when the "create_index" option
on the ids setting is set to "always".

This commit introduces a new enum "flex_table_index_type" which is used
instead of osmium::item_type to define the index type, because
osmium::item_type has no value for "tile".