Better logic for always_build_id_index() check
The logic here should be: We either always want to build an index or we
want to because our database is updatable. And we only do that if there
is an actual id column to build the index on.

The old logic looked a bit different. It was still technically okay,
because always_build_id_index() was never true unless has_id_column() is
also true, but the new logic looks cleaner.