Call middle::after_nodes/ways/relations() in tests in right order
The middle should always be called with these functions in the right
order if the middle data is changed in any way, because there might be
postprocessing done or buffers flushed etc.

This adds some code that checks in the middle that it is in the correct
state. This code is only compiled in in debug mode.