Fix: Fix threading issue in logger
The (singleton) logger can be called from multiple threads at once, but
it writes the m_needs_leading_return member variable. This commit turns
that variable into an atomic to fix this.

This also moves some logic out of the generate_common_prefix() function
so that we can make it const. This makes it easier to see that the code
is now thread-safe.