Disable sequential scan on database tables in copy threads
This forces the DELETEs to use an index scan which is much faster.

The underlying problem here is probably that the DELETEs are called for
all OSM objects in the input file, regardless of whether they are even
in the table. For most tables most ids will not be in there. So
PostgreSQL overestimates the number of "hits" it is going to get and
thinks a full table scan is more efficient when in fact an index lookup
is much better.

The COPYs used in the same database connection are not affected by this.

In many tests I have done I only see this problem in some cases, but
when it hits it can slow down processing considerably. This could
explain the occasional slow processing that some people have been
reporting over the years and that we could never really nail down
(#1674, #1971).