Find node dependencies faster by doing a single lookup
For each changed node osm2pgsql has to find all ways and all relations
referencing that node and make sure they are updated as well. This was
done by doing one SQL query per node which is quite inefficient. Instead
the new code collects all node ids and does only one query for the
parent ways and one for the parent relations after all nodes have been
read.

This implementation uses temporary tables. The changed nodes are written
to a temp table and from that temp tables are filled with the parent
nodes and ways. Strictly speaking the latter tables are not necessary,
we could read the data directly from the query. But maybe those tables
could be useful in the future for some additional processing in the
database. (They'd need to be converted to non-temp tables then, though.)