Don't do multi-statement queries
Only send one SQL command per exec(). This should fix a problem with the
Pgpool-II connection pool server which doesn't support multi-statement
queries. It also makes it hopefully a bit more obvious that there are
several statements sent to the server. To that end I have also removed
the semicolons at the end of the queries in all places, because they
should never be needed any more.

Fixes #1994