Set default schema to 'public' and use qualified names everywhere
Until now the default schema was empty and the PostgreSQL search path
was used to find tables etc.

We now set the default schema for the middle and the pgsql and flex
outputs to 'public'.

This simplifies the code, because we don't have to handle the case
without schema specially any more. More importantly it makes it easier
for the user to reason about what osm2pgsql is doing, because it does
not depend on the setting of the search path in the PostgreSQL session.

This fixes a problem where osm2pgsql could not find the
osm2pgsql_properties table it created itself, because it was created in
a different schema than expected (#2010).

The behaviour of the gazetteer output has not changed. It still uses
unqualified names.

This will be a breaking change for those users who rely on the old
behaviour.

Fixes #2010, #2011.