Fix a crash when an exception is thrown in postprocessing
Postprocessing (like building indexes) on output tables happens in
parallel. If there is a problem an exception is thrown which is
forwarded to the main thread. Here the normal exception handling will
clean up everything including the table datastructures still used by the
postprocessing of the other tables which might still be in flight.

We fix this by storing the exception, waiting for all threads to finish
and only then passing on the exception.

It is rather unlikely that we'll see this problem in the pgsql output
because tables are more or less hardcoded there. It is more likely in
the flex output, where users have many options of creating some havoc
with unusual table definitions or so. In the flex output we store a
pointer to the table so that we can print its name to help narrowing
down the problem somewhat.