Refactor tests triggering a moved-from-warning
Clang-tidy doesn't understand that only one SECTION is run per test so
it thinks data is used after move.

This removes some of the SECTIONs which don't really do things relevant
to the tests here. And one test is split into two so we don't need the
SECTION any more.