Disable the warning that cache is disabled in append mode
In append mode we don't necessarily have the flat node setting from the
command line, because it is read from the properties file. So this would
give us an unnecessary and wrong warning.

This is just a band aid though, a better fix would be to check after we
have read the properties file. But that needs some larger changes I'll
leave for when we refactor the command line parsing code.