Move DISTINCT from queries generating list to when we read the data
We had some problems with the query planner getting confused with those
queries. Lets make them as simple as possible to not confuse it
unnecessarily. This might be vodoo, but it doesn't cost much, just a
little bit of temporary space in the database, because when reading we
do the ORDER BY anyway, so the extra DISTINCT isn't more cost.