Put query for finding parent ways from nodes inside a PL/pgSQL function
The query we had before would sometimes result in really bad performance
when the query planner in PostgreQL decides that it needs to use
sequential scans on the ways table. This is due to the statistics being
way off. By running the query for each node ids on its own we hopefully
trick the query planner into always using an index lookup.

See #2056