Generalization: Allow setting condition for running SQL commands
If a condition is set with "if_has_rows", the SQL command in
"if_has_rows" will be run and only if that returns any rows, the main
sql command is run. This check happens only in "append" mode, in
"create" mode the command is always run.

To be used with something like "SELECT 1 FROM some_table LIMIT 1" to
check whether the table has any rows (for instance an expire table).